Added Software folder, motion project.
Added software folder and prepared motion control project. As of this
commit, the motion control project is a paired down version of the
StellarisWare library, with the motion project itself a near copy of the
qs-logger demo project.

Support for compilers other than IAR EWARM has been omitted to save space,
but is available in the original StellarisWare library.