Developed support for QEI, made hardware fixes.
Developed basic software support for QEI, gets hardware running, may want an
accessor function or two especially for speed so that the sign of the
velocity is included. While testing hardware found that PF0 was blown out,
likely ESD. Used PD6 as backup, used for PWML for hbridge_right. If want
that pin back, will have to cut trace and jumper another pin over to provide
this function.