Finished coding I2C driver, about to change.
Finished writing driver for I2C hardware, found that overarching ST protocol
to communicate with sensors is more complicated than expected. Will need to
modify code to acommodate multiple starts, internal addresses in part. This
code was not test run as hardware would not have communicated well with it,
except in a select case or few.